requirements.txt
fastapi
uvicorn
vllm
python-dotenv
uvicorn 
python-multipart 
pydantic

